requirements.txt
streamlit
textblob
Pillow
numpy
emoji
transformers
torch